[Test] Nav Context 문맥 적용
