+++ package.json
@@ -8,7 +8,7 @@
     "start": "next start",
     "lint": "next lint",
     "stylelint": "stylelint **/*.scss --fix",
-    "test": "jest src/components/layout nav --watch"
+    "test": "jest --watch"
   },
   "dependencies": {
     "@notionhq/client": "^2.1.1",
